rag-example/backend/simulation-function/requirements.txt
functions-framework==3.8.3
Flask==3.1.1
google-genai==1.25.0
google-auth==2.40.3
